LayoutTests/platform/gtk/imported/w3c/web-platform-tests/mathml/relations/css-styling/default-properties-on-semantics-and-maction-expected.txt


FAIL Display value of children of the <semantics> element assert_equals: Child undefined has display: none expected "none" but got "block"
FAIL Display value of children of the <maction> element assert_equals: Child undefined has display: none expected "none" but got "block"
1
1
